requirements.txt
cffi==1.15.1
cryptography==41.0.3
pycparser==2.21
PyJWT==2.8.0
